feat(proteins): create 'proteins' app and move protein model into it
- Created new Django app named 'proteins' to isolate domain logic related to protein data
- Moved the Protein model from the catalog app to 'proteins/models.py'
